pool: Decompose the pool and balancer crates
* Pool::update_pool is replaced with a narrower API:
  * Avoids need for unneeded allocations;
  * Slightly modifies the p2c endpoint update metrics, removing the DNE
    label.
* Split up crates:
  * Move the pool trait to a minimal crate
  * Move the balancer's P2cPool types into a library crate
* Rename NewBalancePeakEwma to NewBalance
