+++ linkerd/pool/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "linkerd-pool"
+version = "0.1.0"
+license = "Apache-2.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+tower-service = "0.3"
