+++ linkerd/proxy/http/src/balance.rs
@@ -3,5 +3,5 @@ pub use linkerd_proxy_balance::*;
 
 pub type Body<B> = PendingUntilFirstDataBody<peak_ewma::Handle, B>;
 
-pub type NewBalancePeakEwma<B, X, R, N> =
-    linkerd_proxy_balance::NewBalancePeakEwma<PendingUntilFirstData, http::Request<B>, X, R, N>;
+pub type NewBalance<B, X, R, N> =
+    linkerd_proxy_balance::NewBalance<PendingUntilFirstData, http::Request<B>, X, R, N>;
