+++ linkerd/proxy/tcp/src/balance.rs
@@ -1,5 +1,5 @@
 pub use linkerd_proxy_balance::*;
 pub use tower::load::CompleteOnResponse;
 
-pub type NewBalancePeakEwma<Req, X, R, N> =
-    linkerd_proxy_balance::NewBalancePeakEwma<CompleteOnResponse, Req, X, R, N>;
+pub type NewBalance<Req, X, R, N> =
+    linkerd_proxy_balance::NewBalance<CompleteOnResponse, Req, X, R, N>;
